+++ src/core/client/views/hud/html/app.js
@@ -187,6 +187,31 @@ const app = new Vue({
             if ('alt' in window) {
                 alt.emit('mouse:Focus', this.leaderboard);
             }
+        },
+        audio3D(soundName, pan, volume, duration = -1) {
+            const audio = new Audio(`./sounds/${soundName}.mp3`);
+            const ambientContext = new AudioContext();
+            const source = ambientContext.createMediaElementSource(audio);
+            const ambientPan = ambientContext.createStereoPanner();
+            source.connect(ambientPan);
+            ambientPan.connect(ambientContext.destination);
+            ambientPan.pan.value = pan;
+
+            if (duration >= 0) {
+                audio.loop = true;
+            } else {
+                audio.loop = false;
+            }
+
+            audio.volume = volume;
+            audio.autoplay = true;
+            audio.play();
+
+            if (duration >= 0) {
+                setTimeout(() => {
+                    audio.pause();
+                }, duration);
+            }
         }
     },
     directives: {
@@ -248,6 +273,7 @@ const app = new Vue({
             alt.on('chat:Focus', this.focusChat);
             alt.on('chat:Inject', this.inject);
             alt.on('leaderboard:Toggle', this.toggleLeaderboard);
+            alt.on('hud:Audio3D', this.audio3D);
             alt.emit('chat:Inject');
         } else {
             let count = 0;